Remove command line interface options
When adding or removing interfaces., dhcp6c would need to be shut down as the interfaces were part of the command line. The interfaces are also stored in the config file. Using the config file rather than the command line means dhcp6c no longer needs to be shutdown, a sighup causes a reset and re-read of the config.