Remove command line interface options, Add a new ENV var - PINFO
When adding or removing interfaces., dhcp6c would need to be shut down as the interfaces were part of the command line. The interfaces are also stored in the config file. Using the config file rather than the command line means dhcp6c no longer needs to be shutdown, a sighup causes a reset and re-read of the config.

PINFO envar holds the allocated PD info assigned by the ISP in the form
ipv6address/len. This can then be used to display on the GUI.