PDINFO - All Interfaces
PDINFO is great, but if you have multiple LAN interfaces tracking it was only showing one of them. This update now puts the interface name and the prefix information for every interface that is tracking. It does so in the format of:

interface.prefix, so for example igb1.2a02:xxxx:xxxxx:xx.../64 igb2.2a02:xxxx:xxxxx:xx.../64
etc

Added null terminator space

i