config: shorten the code
Multiple ifid statements will be accepted.  The last one wins.
This is the same as ifid-random, either one of the last one wins.

To be honest I'd rather have liked to see

    ifid random;

This is a bit more intuitive from the user side and allows to
make the override of the override to make more sense as it is
still the same statement.