parser: allow "ifid random;" syntax
Might be overkill at this point, but if we add more non-numeric
modes like a sensible fallback for PPPoE which doesn't do EUI-64
and the daemon dies as a result this seems easier to explain.

One note here is that the ifid is random per config parse and
since we allow SIGHUP reload we get a random address each time.