config: negative numbers are not supported, use strtoull
The lexer won't recognizea "-" so strtoll() lower limit (highest
bit) can never be used.  For the ifid this is a problem as it
would cap the user input as it can be a full 64 bit wide number
but would not reflect it in the resulting address.

So we switch to strtoull() and assume that most spots have sanity
checks for upper bounds and the pieces that use a -1 for "infinity"
users need to understand that setting the max unsigned long long
value they will have reached infinity indeed.  ;)