dhcp6c: fix release/remove madness
free_resources() is the final operation so it doesn't make sense to
keep a RELEASE event here, especially since we call remove_ia() which
removes further means of reacting to the event.  It also tries to INIT
the client again before we parse the new configuration so that is quite
counter-productive.

PR: https://forum.opnsense.org/index.php?topic=41950.0