common: fix underflow, enforce RAND boundaries correctly
So the problem was that arc4random*() is unsigned int, but
random() was signed long so "- 2000" would produce an underflow
and broke the result of the random skew value in 50% of the
cases.

While here enforce the strict boundaries to end up at +-0.1
except in the "strictly greater than zero" case.

And instead of risking an underflow we can also just calculate
the range as 0.0 - 0.2 and do -0.1 when we are in double
precision.